Add Hue logging
Adds logging of state changes, so we can track outcomes.

Light state change failure is ignored by consumers, since this is a fire-and-forget sort of API.

Fixes a bug with light states; the chaining API actually modifies the original object, so we can't just reuse them.

Removes methods that just access the config. It's not the job of the Hue API to provide that, ask the config service.