Move DB management out of Plugin and switch to LevelDB
Removes LokiJS in favor of a LevelDB key-value store. LokiJS, though nice and convenient, felt like overkill with its Mongo-like semantics. I think the types of data that plugins are likely to want to keep track of will match up better with a key-value store.

Reimplements database management to move it out of WatneyPluginBase and into its own interface and provider(s). This reduces WatneyPluginBase's responsibilities significantly and the resulting code is a lot easier to test. Currently, only the LevelDB provider is included and WatneyPluginBase uses it.

TODO:
- Update the CLI for TypeScript
- Rewrite readme