Fix app module path for CLI
The CLI needs to import the app module file in order to get the plugins and discover their CLIs. However, we were still looking for the app module using regular Node conventions, rather than taking into account TypeScript transpilation. Since each app can decide how to do that, we leave it to a package.json setting to point us to the transpiled module.