Fix unconfigured plugin issue
Attempting to construct a plugin (such as when firing up the CLI) when there was no config yet would cause errors like 'Configuration property "hue" is not defined'. This is unacceptable, since it's part of the design that users will run the CLI _in order to_ figure out how to configure the plugin.